requirements.txt
scikit-learn==0.24.2
numpy
pandas
prince
umap-learn
SentenceTransformer